pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pytest-playwright", "pytest-unused-port"]

[tool.setuptools]
packages = []
